BAD_Web/requirements.txt
djangorestframework==3.1.1
Django==2.1.5
mysqlclient
mysql-connector-python==8.0.5
MySQL-python==1.2.5
djangorestframework-jsonapi